commit for task six

+++ 0x00-hello_world/6-size.c
@@ -12,7 +12,7 @@ long int b;
 lon long int c;
 char d;
 float f;
-printf("size of a char: %d byte(s)\n", (unsigned)sizeof(d));
+printf("size of a char: %d byte(s)\n", (unsigned long)sizeof(d));
 printf("size of an int: %d byte(s)\n", (unsigned long)sizeof(a));
 printf("size of a long int: %d byte(s)\n", (unsigned long)sizeof(b));
 printf("size of a long long int: %d byte(s)\n", (unsigned long)sizeof(c));
